+++ Media Blocks SDK/Android/RTSP Client/MainActivity.cs
@@ -6,6 +6,7 @@ using Android.Util;
 using System;
 using VisioForge.Core;
 using VisioForge.Core.MediaBlocks;
+using VisioForge.Core.MediaBlocks.AudioRendering;
 using VisioForge.Core.MediaBlocks.Sources;
 using VisioForge.Core.MediaBlocks.VideoRendering;
 using VisioForge.Core.Types.Events;
@@ -18,7 +19,7 @@ namespace RTSP_Client
     {
         private string TEST_URL = "rtsp://";
 
-        private VisioForge.Core.UI.Android.VideoView videoView;
+        private VisioForge.Core.UI.Android.VideoViewTX videoView;
 
         private Button btStart;
 
@@ -32,36 +33,27 @@ namespace RTSP_Client
 
         private EditText edPassword;
 
-        private GridLayout pnScreen;
-
         //private readonly System.Timers.Timer tmPosition = new System.Timers.Timer(500);
-
-        private bool isSeeking = false;
 
         private MediaBlocksPipeline _pipeline;
 
         private VideoRendererBlock _videoRenderer;
 
-        //private AudioRendererBlock _audioRenderer;
+        private AudioRendererBlock _audioRenderer;
 
         private RTSPSourceBlock _source;
 
-        private void DestroyEngine()
+        private async Task DestroyEngineAsync()
         {
             if (_pipeline != null)
             {
                 _pipeline.OnError -= _pipeline_OnError;
-                // _pipeline.OnStop -= _pipeline_OnStop;
-                //_pipeline.OnStart -= _pipeline_OnStart;
 
-                _pipeline.Dispose();
+                await _pipeline.DisposeAsync();
                 _pipeline = null;
             }
         }
 
-        //private async void _pipeline_OnStart(object sender, EventArgs e)
-        //{
-        //}
 
         private void _pipeline_OnError(object sender, ErrorsEventArgs e)
         {
@@ -75,10 +67,6 @@ namespace RTSP_Client
             base.OnDestroy();
         }
 
-        //private void _pipeline_OnStop(object sender, StopEventArgs e)
-        //{
-        //}
-
         protected override void OnCreate(Bundle? savedInstanceState)
         {
             base.OnCreate(savedInstanceState);
@@ -88,22 +76,12 @@ namespace RTSP_Client
 
             RequestPermissions(
                 new String[]{
-                        Manifest.Permission.Camera,
                         Manifest.Permission.Internet,
-                        Manifest.Permission.RecordAudio,
                         Manifest.Permission.AccessNetworkState,
-                        Manifest.Permission.AccessWifiState,
-                        Manifest.Permission.ModifyAudioSettings}, 1004);
+                        Manifest.Permission.AccessWifiState},
+                1004);
 
-             videoView = FindViewById<VisioForge.Core.UI.Android.VideoView>(RTSP_Client.Resource.Id.videoView);
-
-            _pipeline = new MediaBlocksPipeline();
-            //_pipeline.Debug_Dir = System.IO.Path.Combine(System.Environment.GetFolderPath(System.Environment.SpecialFolder.MyDocuments), "VisioForge");
-            _pipeline.OnError += _pipeline_OnError;
-            // _pipeline.OnStop += _pipeline_OnStop;
-            // _pipeline.OnStart += _pipeline_OnStart;
-
-            _videoRenderer = new VideoRendererBlock(_pipeline, videoView) { IsSync = false };
+            videoView = FindViewById<VisioForge.Core.UI.Android.VideoViewTX>(RTSP_Client.Resource.Id.videoView);
 
             btStart = FindViewById<Button>(RTSP_Client.Resource.Id.btStart);
             btStart.Click += btStart_Click;
@@ -113,8 +91,6 @@ namespace RTSP_Client
 
             btStop = FindViewById<Button>(RTSP_Client.Resource.Id.btStop);
             btStop.Click += btStop_Click;
-
-            pnScreen = FindViewById<GridLayout>(RTSP_Client.Resource.Id.pnScreen);
 
             edURL = FindViewById<EditText>(RTSP_Client.Resource.Id.edURL);
             edURL.Text = TEST_URL;
@@ -137,11 +113,7 @@ namespace RTSP_Client
 
             videoView.Invalidate();
 
-            //DestroyEngine();
-
-            // clear screen workaround
-            pnScreen.RemoveView(videoView);
-            pnScreen.AddView(videoView);
+            await DestroyEngineAsync();
         }
 
         private async void btPause_Click(object sender, EventArgs e)
@@ -165,67 +137,27 @@ namespace RTSP_Client
 
         private async void btStart_Click(object sender, EventArgs e)
         {
-            //var src = new VirtualVideoSourceBlock(new VirtualVideoSourceSettings());
-            //var src = new AndroidVideoSourceBlock(new AndroidVideoSourceSettings());
+            _pipeline = new MediaBlocksPipeline();
+            _pipeline.OnError += _pipeline_OnError;
 
-            isSeeking = false;
-
-            //var mediaInfo = new MediaInfoReaderX(context: null, this);
-            //bool videoStream = true;
-            //bool audioStream = true;
-            //if (await mediaInfo.OpenAsync(new Uri(edURL.Text)))
-            //{
-            //    if (mediaInfo.Info.VideoStreams.Count == 0)
-            //    {
-            //        videoStream = false;
-            //    }
-
-            //    if (mediaInfo.Info.AudioStreams.Count == 0)
-            //    {
-            //        audioStream = false;
-            //    }
-            //}
-
-            //var data = System.IO.File.ReadAllBytes(edURL.Text);
-            //_memorySource = new StreamSourceBlock(new MemorySourceSettings(data));
-
-            //_decodeBin = new DecodeBinBlock(true, true, false);
-            //_pipeline.Connect(_memorySource.Output, _decodeBin.Input);
-
-            var rtspSource = await RTSPSourceSettings.CreateAsync(new System.Uri(edURL.Text), "admin", "password", audioEnabled: true);
+            var rtspSource = await RTSPSourceSettings.CreateAsync(new System.Uri(edURL.Text), edLogin.Text, edPassword.Text, audioEnabled: true);
 
             _source = new RTSPSourceBlock(rtspSource);
 
-            // if (videoStream)
-            //{
-            _pipeline.Connect(_source.VideoOutput, _videoRenderer.Input);
-            // }
+            if (_source.Settings.IsVideoAvailable())
+            {
+                _videoRenderer = new VideoRendererBlock(_pipeline, videoView) { IsSync = false };
+                _pipeline.Connect(_source.VideoOutput, _videoRenderer.Input);
+            }
 
-            //if (audioStream)
-            // {
-            // _audioRenderer = new AudioRendererBlock();
-            // _pipeline.Connect(_source.AudioOutput, _audioRenderer.Input);
-            // }
+            if (_source.Settings.IsAudioAvailable())
+            {
+                _audioRenderer = new AudioRendererBlock();
+                _pipeline.Connect(_source.AudioOutput, _audioRenderer.Input);
+            }
 
             await _pipeline.StartAsync();
         }
-
-        //public override bool OnCreateOptionsMenu(IMenu menu)
-        //{
-        //    MenuInflater.Inflate(Resource.Menu.menu_main, menu);
-        //    return true;
-        //}
-
-        //public override bool OnOptionsItemSelected(IMenuItem item)
-        //{
-        //    int id = item.ItemId;
-        //    if (id == Resource.Id.action_settings)
-        //    {
-        //        return true;
-        //    }
-
-        //    return base.OnOptionsItemSelected(item);
-        //}
 
         public override void OnRequestPermissionsResult(int requestCode, string[] permissions, [GeneratedEnum] Android.Content.PM.Permission[] grantResults)
         {